+++ src/v2/run-es.js
@@ -1,8 +1,7 @@
 const semver = require('semver');
-const waitOn = require('wait-port');
 const esUtils = require('./es-utils');
 const kbUtils = require('./kb-utils');
-const logger = require('../log')(module.id);
+const logger = require('../log')(module);
 
 /* eslint no-sync: 0 */
 class Elastic {
@@ -16,99 +15,129 @@ class Elastic {
     this.app.locals.env.ES_REASON = [];
     this.app.locals.env.ES_ACTIVE = false;
     if (!this.app.locals.env.ES_HOST) {
-      this.app.locals.env.ES_REASON.push(`No value for ES_HOST variable in config!`);
-      this.app.logger.log('warn', `[INIT] - No value for ES_HOST variable in config - data will NOT be saved!`);
-    }
-    try {
-      this.waitLoop(attempt, attempts);
-    } catch (e) {
-      this.app.locals.env.ES_REASON.push(`Error setting up Elastic! Data will NOT be saved to ES! Error: ${e}`);
-      logger.log('error', `[INIT] Error setting up Elastic! Data will NOT be saved to ES!`, e);
+      this.app.locals.env.ES_REASON.push(`[INIT] value for ES_HOST variable in config: [false]`);
+      this.app.logger.log('warn', `[INIT] No value for ES_HOST variable in config - data will NOT be saved!`);
+    } else {
+      try {
+        this.waitLoop(attempt, attempts);
+      } catch (e) {
+        this.app.locals.env.ES_REASON.push(`[INIT] Error setting up Elastic! Data will be saved to ES: [false]! Error: ${e}`);
+        logger.log('error', `[INIT] Error setting up Elastic! Data will NOT be saved to ES!`, e);
+      }
     }
   }
 
   waitLoop(attempt = 1, attempts = 5) {
-    let waitResults;
-    logger.log('info', `[INIT] Waiting for Elasticsearch and Kibana - attempt #${attempt} ...`);
-    setTimeout(async () => {
-      waitResults = await this.waitES();
-      if (attempt < attempts && !waitResults) {
-        this.waitLoop(attempt + 1);
-      } else if (!waitResults) {
-        const msg = `No response from Elasticsearch and/or Kibana - are they running? Please check ` +
-          `[${this.app.locals.env.ES_HOST}:${this.app.locals.env.ES_PORT}]`;
-        this.app.locals.env.ES_REASON.push(msg);
-        logger.log('warn', `[INIT] ${msg}`);
-      } else {
-        logger.log('info', `[INIT] Elasticsearch and Kibana are available!`);
-      }
-    }, this.app.locals.env.ES_TIMEOUT);
+    logger.log('info', `[INIT] Checking status of Kibana and Elasticsearch [attempt # ${attempt} of 5]`);
+    this.waitES()
+      .then(waitResults => {
+        if (!waitResults) {
+          if (attempt < attempts) {
+            logger.log('info', `[INIT] Waiting ${this.app.locals.env.ES_TIMEOUT / 1000} seconds to check status of ` +
+              `Kibana and Elasticsearch again ...`);
+            setTimeout(async () => {
+              if (attempt < attempts && !waitResults) {
+                waitResults = this.waitLoop(attempt + 1);
+              } else if (!waitResults) {
+                this.app.locals.env.ES_REASON.push(
+                  `[INIT] Response from Elasticsearch and or Kibana: [false] - are they running? Please check ` +
+                    `[${this.app.locals.env.ES_HOST}:${this.app.locals.env.ES_PORT}]`);
+              }
+            }, this.app.locals.env.ES_TIMEOUT);
+          } else {
+            logger.log('warn', 'Could not connect to Kibana or Elasticsearch!\nPlease check your config and/or ' +
+              'check the health of the elastic cluster!\nThe Timings API is running at ' +
+              `[http://${this.app.locals.env.HOST}:${this.app.locals.env.HTTP_PORT}] but data will NOT be saved to Elasticsearch`
+            );
+          }
+        } else {
+          logger.log('info', this.app.locals.env.ES_REASON);
+          if (this.app.locals.env.ES_ACTIVE === true) {
+            logger.log('info', `\n${'='.repeat(100)}\nThe Timings API is running at ` +
+              `[http://${this.app.locals.env.HOST}:${this.app.locals.env.HTTP_PORT}]\n${'='.repeat(100)}`
+            );
+          }
+        }
+      });
   }
 
   async waitES() {
-    // Wait for healthy status and get info
-    const waitESResult = await waitOn({
-      host: this.app.locals.env.ES_HOST,
-      port: this.app.locals.env.ES_PORT,
-      output: 'silent',
-      timeout: this.app.locals.env.ES_TIMEOUT || 5000
-    });
-    const waitKBResult = await waitOn({
-      host: this.app.locals.env.KB_HOST,
-      port: this.app.locals.env.KB_PORT,
-      output: 'silent',
-      timeout: this.app.locals.env.ES_TIMEOUT || 5000
-    });
-    if (waitKBResult) await this.checkKBStatus();
-    if (waitESResult) await this.checkESStatus();
-    return waitESResult && waitKBResult;
+    const kbStatus = await this.checkKBStatus();
+    if (kbStatus?.status?.overall?.state === 'green') {
+      await this.checkESStatus();
+      if (this.app.locals.env.ES_ACTIVE) {
+        if (this.app.locals.esUpgrade === true) {
+          // Forced upgrade --es-upgrade flag is set
+          await this.upgrade();
+          return kbStatus;
+        }
+        // Check for necessary ES upgrades
+        const upgrades = await this.checkUpgrade();
+        if (upgrades.type && typeof this[upgrades.type] === 'function') {
+          const response = this[upgrades.type]();
+          if (!response.ok) {
+            this.app.locals.env.ES_REASON.push(
+              `[ELASTIC] Upgrade failed ... Elasticsearch will NOT be active and data will not be saved!`
+            );
+            this.app.locals.env.ES_ACTIVE = false;
+          }
+        }
+      }
+    }
+    return kbStatus?.status?.overall?.state === 'green';
   }
 
   async checkKBStatus() {
     const healthResult = await this.kb.getKBStatus();
-    if (healthResult.status) {
-      this.app.locals.env.KB_VERSION = healthResult.version.number || 'unknown';
-      this.app.locals.env.KB_BUILD = healthResult.version.build_number || 'unknown';
-      this.app.locals.env.KB_STATUS = healthResult.status.overall?.state || 'unknown';
-      this.app.locals.env.ES_REASON.push(`Kibana status: ${this.app.locals.env.KB_STATUS}`);
-      const esStatus = healthResult.status.statuses.filter(s => s.id.indexOf('core:elasticsearch') === 0);
+    this.app.locals.env.KB_VERSION = healthResult?.version?.number || 'unknown';
+    this.app.locals.env.KB_BUILD = healthResult?.version?.build_number || 'unknown';
+    this.app.locals.env.KB_STATUS = healthResult?.status?.overall?.state || 'unknown';
+    this.app.locals.env.KB_INDEX_VERSION = healthResult?.version?.number || '0.0.0';
+    this.app.locals.env.KB_INDEX_MAJOR = parseInt(this.app.locals.env.KB_INDEX_VERSION.substring(0, 1), 10);
+    if (Array.isArray(healthResult?.status?.statuses)) {
+      const esStatus = healthResult?.status?.statuses.filter(s => s.id.indexOf('core:elasticsearch') === 0);
       if (esStatus.length > 0) {
         this.app.locals.env.ES_STATUS = esStatus[0].state || 'unknown';
         if (['green', 'yellow'].includes(this.app.locals.env.ES_STATUS)) {
           this.app.locals.env.ES_ACTIVE = true;
-          this.app.locals.env.ES_REASON.push(`Elasticsearch  status: ${this.app.locals.env.ES_STATUS}`);
+          this.app.locals.env.ES_REASON.push(
+            `[ELASTIC] Elasticsearch status good: [${['green', 'yellow'].includes(this.app.locals.env.ES_STATUS)}]`);
         }
       }
-    }
+    } else this.app.locals.env.ES_STATUS = 'unknown';
 
     // const healthResult = await this.es.healthy(`${this.app.locals.env.ES_TIMEOUT / 10}s`, 'yellow');
-    if (!this.app.locals.env.KB_STATUS === 'green') {
-      this.app.locals.env.ES_REASON.push(`Kibana state is [${this.app.locals.env.KB_STATUS}] while expecting [green] - ` +
-        `Please check [${this.app.locals.env.ES_PROTOCOL}://${this.app.locals.env.KB_HOST}:${this.app.locals.env.KB_PORT}]`);
+    if (this.app.locals.env.KB_STATUS === 'green') {
+      this.app.locals.env.ES_REASON.push(`[KIBANA] Kibana status green: [true]`);
     } else {
-      // Elastic looks good!
-      this.app.locals.env.ES_ACTIVE = true;
+      this.app.locals.env.ES_REASON.push(`[KIBANA] Kibana status green: [false] - ` +
+        `Please check [${this.app.locals.env.ES_PROTOCOL}://${this.app.locals.env.KB_HOST}:${this.app.locals.env.KB_PORT}]`);
     }
+    return healthResult;
   }
 
   async checkESStatus() {
     if (['green', 'yellow'].includes(this.app.locals.env.ES_STATUS)) {
       this.app.locals.env.ES_ACTIVE = true;
-      await this.checkEsTemplate();
       const info = await this.es.info();
       const version = info?.version?.number;
       if (!version) {
-        this.app.locals.env.ES_REASON.push(`No version found for Elasticsearch - does your version match the client? ` +
-          `Please check [${this.app.locals.env.ES_PROTOCOL}://${this.app.locals.env.ES_HOST}:${this.app.locals.env.ES_PORT}]`);
+        this.app.locals.env.ES_REASON.push(
+          `[ELASTIC] Version found for Elasticsearch: [false] - does your version match the client? ` +
+          `Please check [${this.app.locals.env.ES_PROTOCOL}://${this.app.locals.env.ES_HOST}:${this.app.locals.env.ES_PORT}]`
+        );
         this.app.locals.env.ES_ACTIVE = false;
       } else {
+        const templateVersion = await this.es.getTemplateVersion();
+        this.app.locals.env.CURR_VERSION = semver.valid(templateVersion) || null;
+        this.app.locals.env.NEW_VERSION = semver.valid(this.app.locals.env.APP_VERSION) || this.app.locals.env.CURR_VERSION;
         this.app.locals.env.ES_VERSION = version;
         this.app.locals.env.ES_VERSION_INFO = info.version;
         this.app.locals.env.ES_CLUSTER = info.cluster_name;
         this.app.locals.env.ES_MAJOR = parseInt(version.substring(0, 1), 10);
         this.app.locals.env.ES_ACTIVE = this.es.clientVersionMajor === this.app.locals.env.ES_MAJOR;
-        this.app.locals.env.ES_REASON.push(`NPM client [v.${this.es.clientVersion}] and ` +
-          `Elasticsearch [v.${this.app.locals.env.ES_VERSION}] - are ${this.app.locals.env.ES_ACTIVE ? '' : 'NOT '}a match!`);
+        this.app.locals.env.ES_REASON.push(`[ELASTIC] NPM client [v.${this.es.clientVersion}] and ` +
+        `Elasticsearch [v.${this.app.locals.env.ES_VERSION}] - match: [${this.app.locals.env.ES_ACTIVE}]`);
       }
     }
   }
@@ -183,7 +212,8 @@ class Elastic {
     } else {
       response = {
         ok: false,
-        info: [`Elasticsearch is not in use! Reason(s): ${this.app.locals.env.ES_REASON}`]
+        info: `Elasticsearch is not in use!`,
+        reason: this.app.locals.env.ES_REASON
       };
     }
     return response;
@@ -213,7 +243,7 @@ class Elastic {
       info: [],
       ok: false
     };
-
+    // There is reason to upgrade but the '--es-upgrade' is not set! Skip upgrade and set ES to inactive!
     if (this.app.locals.env.ES_ACTIVE === true) {
       const esTemplate = require('../../config/.es_template.json');
       esTemplate.template.mappings._meta.api_version = this.app.locals.env.APP_VERSION;
@@ -227,16 +257,16 @@ class Elastic {
       response.kibanaStatus = this.app.locals.env.KB_STATUS;
       response.kibanaIndex = this.app.locals.env.KB_INDEX;
       try {
-        logger.log('info', `[UPDATE] updating index template [${this.app.locals.env.INDEX_PERF}] ...`);
+        logger.log('info', `[UPGRADE] updating index template [${this.app.locals.env.INDEX_PERF}] ...`);
         await this.es.putTemplate(this.app.locals.env.INDEX_PERF, esTemplate);
-        logger.log('info', `[UPDATE] index template [${this.app.locals.env.INDEX_PERF}] updated successfully!`);
+        logger.log('info', `[UPGRADE] index template [${this.app.locals.env.INDEX_PERF}] updated successfully!`);
         response.ok = true;
       } catch (e) {
         response.ok = false;
         response.err = e;
       }
     } else {
-      response.info.push(
+      logger.log('warn',
         `Elasticsearch is not in use - cannot upgrade template [${this.app.locals.env.INDEX_PERF}*]`);
     }
     return response;
@@ -244,44 +274,39 @@ class Elastic {
 
   async checkUpgrade() {
     const response = {
-      action: 'Template version check',
-      info: []
+      action: 'Upgrade check',
+      info: [],
+      upgrade: false
     };
 
     if (this.app.locals.env.ES_ACTIVE === true) {
-      this.app.locals.env.NEW_VERSION = semver.valid(this.app.locals.env.APP_VERSION) || this.app.locals.env.CURR_VERSION;
-      this.app.locals.env.KB_INDEX_VERSION = await this.kb.getKBVer();
-      this.app.locals.env.KB_INDEX_MAJOR = parseInt(this.app.locals.env.KB_INDEX_VERSION.substring(0, 1), 10);
+      // Check the API version that created the current 'cicd-perf' template (if it's present)
       response.currVersion = this.app.locals.env.CURR_VERSION;
       response.newVersion = this.app.locals.env.NEW_VERSION;
-
-      const upgrade = (
-        this.app.locals.env.CURR_VERSION && (
-          semver.lt(this.app.locals.env.CURR_VERSION, this.app.locals.env.NEW_VERSION) ||
-          (this.app.locals.env.KB_INDEX_MAJOR < this.app.locals.env.ES_MAJOR) ||
-          this.app.locals.ES_UPGRADE === true
-        )
-      );
-      if (upgrade === true) {
-        response.info.push(`[UPDATE] Force: ${this.app.locals.ES_UPGRADE || false} - ` +
-          `App Update: ${semver.lt(this.app.locals.env.CURR_VERSION, this.app.locals.env.NEW_VERSION)} - ` +
-          `(CURR:${this.app.locals.env.CURR_VERSION} / NEW:${this.app.locals.env.NEW_VERSION}) - ` +
-          `ES Upgrade: ${this.app.locals.env.KB_INDEX_MAJOR < this.app.locals.env.ES_MAJOR} - ` +
-          `(KB index major:${this.app.locals.env.KB_INDEX_MAJOR} / ES major:${this.app.locals.env.ES_MAJOR})`);
-        logger.log('info', response.info);
-
-        // Update the [cicd-perf] template
-        this.upgrade();
-
-      } else {
-        response.info.push(
-          `Your Elasticsearch template [${this.app.locals.env.INDEX_PERF}*] was created by the current version of Timings API `
-          + `(v.${this.app.locals.env.APP_VERSION})! No changes needed.`);
+      // Check if this is a first-time install - if so, create the 'cicd-perf' template and set the default index-pattern
+      if (!this.app.locals.env.CURR_VERSION) {
+        logger.log('info', `[ELASTIC] New install - installing templates and sample data`);
+        response.upgrade = true;
+        response.type = 'newInstall';
+      } else if (
+        semver.lt(this.app.locals.env.CURR_VERSION, this.app.locals.env.NEW_VERSION) ||
+        this.app.locals.env.KB_INDEX_MAJOR < this.app.locals.env.ES_MAJOR
+      ) {
+        this.app.locals.env.ES_REASON.push(
+          `[UPGRADE] API version: Current: [${this.app.locals.env.CURR_VERSION}] - New: [${this.app.locals.env.NEW_VERSION}] - ` +
+            `match: [${semver.eq(this.app.locals.env.CURR_VERSION, this.app.locals.env.NEW_VERSION)}]`);
+        this.app.locals.env.ES_REASON.push(
+          `[UPGRADE] KB major version: ${this.app.locals.env.KB_INDEX_MAJOR} - ` +
+            `ES major version: ${this.app.locals.env.ES_MAJOR} - ` +
+            `match: [${this.app.locals.env.KB_INDEX_MAJOR === this.app.locals.env.ES_MAJOR}]`);
+        response.upgrade = true;
+        response.type = 'upgrade';
       }
     } else {
-      response.info.push(
-        `Elasticsearch is not in use - cannot check template [${this.app.locals.env.INDEX_PERF}*]`);
+      this.app.locals.env.ES_REASON.push(
+        `[ELASTIC] Elasticsearch is not in use - checking for upgrades: [false]`);
     }
+    response.info = this.app.locals.env.ES_REASON;
     return response;
   }
 }